fix: compiling errors
