+++ apps/web/components/CategoryItemsPerPage/CategoryItemsPerPage.vue
@@ -5,7 +5,12 @@
     </h6>
 
     <div class="px-4">
-      <SfSelect v-model="selected" :aria-label="$t('perPage')" @change="updateItemsPerPage(Number(selected))">
+      <SfSelect
+        v-model="selected"
+        :aria-label="$t('perPage')"
+        id="perPage"
+        @change="updateItemsPerPage(Number(selected))"
+      >
         <option v-for="{ value, label, disabled } in options" :key="value" :value="value" :disabled="disabled">
           {{ label }}
         </option>
